Add files via upload
This a Chakra game all made with html. Please choose you chakra petal of who you are and check out the gif image to inspire you!. You need to use javaScript for a back end result forms. On it on the next upcoming lessons we will keep updating. To all of you love and peace.